requirements.txt
# Python 3.9.13
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-environ==0.9.0
django-ninja==0.20.0
idna==3.4
pydantic==1.10.4
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
